scripts/pyproject.toml
[project]
name = "scripts"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12.11"
dependencies = [
    "anthropic==0.25.0",
    "beautifulsoup4==4.12.2",
    "discord-py==2.3.2",
    "lxml==4.9.3",
    "python-dotenv==1.0.0",
    "requests==2.31.0",
    "supabase==2.4.0",
]
